requirements.txt
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
confluent-kafka>=2.3.0

# Deep Learning
tensorflow>=2.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Web API
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0

# Excel for EIA fallback (xls + xlsx)
openpyxl>=3.1.0
xlrd>=2.0.0

# Time Series Forecasting
statsmodels>=0.14.0
prophet>=1.1.5

# Transformer NLP (optional: fine-tuned DistilBERT, T5/DistilBART summarization)
torch>=2.0.0
transformers>=4.30.0

# Knowledge Graph (Neo4j)
neo4j>=5.14.0